feat: 完成登录功能 JWT 鉴权 + 添加 .gitignore 和 Spring Profile 配置
- User 实体字段重命名: name→account, username→nickname, userId→Long
- 实现 /api/auth/login, /api/auth/register, /api/auth/profile
- JWT 过滤器无效 token 返回 401
- 注册密码强度校验 (>= 6位)
- 注册成功后返回 JWT token 实现自动登录
- 修复多处编译错误 (重复 package、@ControllerAdvice 包范围、AlbumPhotoMapper 重复声明)
- 添加 .gitignore 排除编译产物和依赖
- application.yaml 改为 blog (生产配置)
- 添加 application-local.yaml (本地开发用 code2026)
- 修复 AlbumController 缺少 albumName 参数

+++ springboot/src/main/java/com/example/springboot/controller/UserController.java
@@ -1,6 +1,5 @@
 package com.example.springboot.controller;
 
-
 import com.example.springboot.common.Result;
 import com.example.springboot.entity.User;
 import com.example.springboot.servise.UserService;
@@ -15,17 +14,16 @@ import org.springframework.web.bind.annotation.RestController;
 public class UserController {
     @Resource
     UserService userService;
-    //首页显示
+
     @GetMapping("/Manager")
-    public Result managerUserData(@RequestParam Integer userId){
+    public Result managerUserData(@RequestParam Long userId){
         User user=userService.managerPage(userId);
         return Result.success(user);
     }
-    //其他显示
+
     @GetMapping("/header")
-    public Result headerUserData(@RequestParam Integer userId){
+    public Result headerUserData(@RequestParam Long userId){
         User user=userService.headerPage(userId);
         return Result.success(user);
     }
-
 }
